+++ src/importexport/midi/internal/midiexport/exportmidi.cpp
@@ -60,7 +60,7 @@ void ExportMidi::writeHeader()
         Staff* staff  = m_score->staff(staffIdx);
 
         ByteArray partName = staff->partName().toUtf8();
-        int len = partName.size() + 1;
+        size_t len = partName.size() + 1;
         unsigned char* data = new unsigned char[len];
 
         memcpy(data, partName.constData(), len);
@@ -69,7 +69,7 @@ void ExportMidi::writeHeader()
         ev.setType(ME_META);
         ev.setMetaType(META_TRACK_NAME);
         ev.setEData(data);
-        ev.setLen(len);
+        ev.setLen(static_cast<int>(len));
 
         track1.insert(0, ev);
